# authenticate/model/responsemodel/common.go
package responsemodel

import "authenticate/model/enum"

type CommonErrorResponse struct {
	ErrorCode    enum.ErrorCode `json:"error_code"`
	ErrorMessage string         `json:"error_message"`
	Result       interface{}    `json:"result"`
}

type CommonResponse struct {
	Result interface{} `json:"result"`
}